Add very basic check-in/check-out functionality
This basically cheats and gets the image you registered with and uses
that to check you in/out via the API.

Probably incredibly brittle but it works and is an interesting proof of
concept.

It would be better if it knew when you were already checked in/out and
sent messages as appropriate. At the moment it's simply based on however
the backend works it out.